refactor: Change the 'common' package structure
