--- src/main/java/com/hexagonal/courseregistration/common/ErrorResponse.java
@@ -1,6 +1,0 @@
-package com.hexagonal.courseregistration.common;
-
-record ErrorResponse(
-  String message
-) {
-}
