--- src/main/java/com/hexagonal/courseregistration/common/SuccessResponse.java
@@ -1,6 +1,0 @@
-package com.hexagonal.courseregistration.common;
-
-record SuccessResponse(
-  String message
-) {
-}
