+++ src/main/java/com/hexagonal/courseregistration/common/adapter/web/ApplicationExceptionHandler.java
@@ -1,11 +1,12 @@
-package com.hexagonal.courseregistration.common;
+package com.hexagonal.courseregistration.common.adapter.web;
 
+import com.hexagonal.courseregistration.common.application.ApiException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
-import static com.hexagonal.courseregistration.common.JsonResponse.badRequest;
-import static com.hexagonal.courseregistration.common.JsonResponse.fail;
+import static com.hexagonal.courseregistration.common.adapter.web.ResponseFactory.badRequest;
+import static com.hexagonal.courseregistration.common.adapter.web.ResponseFactory.fail;
 
 @RestControllerAdvice
 public class ApplicationExceptionHandler {
