+++ src/main/java/com/hexagonal/courseregistration/common/adapter/web/ErrorResponse.java
@@ -1,0 +1,6 @@
+package com.hexagonal.courseregistration.common.adapter.web;
+
+record ErrorResponse(
+  String message
+) {
+}
