+++ src/main/java/com/hexagonal/courseregistration/common/adapter/web/SuccessResponseWithData.java
@@ -1,4 +1,4 @@
-package com.hexagonal.courseregistration.common;
+package com.hexagonal.courseregistration.common.adapter.web;
 
 record SuccessResponseWithData(
   String message,
