+++ src/main/java/com/hexagonal/courseregistration/user/application/ExistUserException.java
@@ -1,6 +1,6 @@
 package com.hexagonal.courseregistration.user.application;
 
-import com.hexagonal.courseregistration.common.ApiException;
+import com.hexagonal.courseregistration.common.application.ApiException;
 import org.springframework.http.HttpStatus;
 
 public class ExistUserException extends ApiException {
